pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telemetry-simulator"
version = "1.0.0"
description = "Schema-driven OTEL telemetry simulator for LLM observability"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = []
keywords = [
    "telemetry",
    "opentelemetry",
    "otlp",
    "simulator",
    "mcp",
    "llm",
    "observability",
    "tracing",
    "metrics",
    "gen-ai",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "black==26.1.0",
    "ruff>=0.15.2",
    "pytest>=9.0.2",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
telemetry-simulator = "simulator.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["simulator*"]

[tool.setuptools.package-data]
"simulator.scenarios.definitions" = ["*.yaml"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "integration: marks tests as integration tests (require external services)",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
# Allow untyped defs until annotations are added; re-enable for stricter checking.
disallow_untyped_defs = false
